Added cli option to choose model revision
Bandit reported an issue about downloading models from huggingface
without a revision number. This change does not fully resolve that issue
as it would require hardcoding revision numbers to a specific commit but
still allows for users (me in the future) to have the choice of
specifying one. Note the bandit issue was ignored so future commits
would not be prevented by CI that will be soon implemented.